backend/auth_service/app/security.py
"""
Auth Service - JWT, Password Hashing, MFA Utilities
"""
import hashlib
import secrets
import uuid
from datetime import datetime, timedelta
from typing import Optional, Dict, Any

import jwt
import pyotp
from passlib.context import CryptContext

from app.config import settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


class PasswordManager:
    @staticmethod
    def hash_password(password: str) -> str:
        return pwd_context.hash(password)

    @staticmethod
    def verify_password(plain: str, hashed: str) -> bool:
        return pwd_context.verify(plain, hashed)

    @staticmethod
    def validate_strength(password: str) -> tuple[bool, str]:
        if len(password) < 8:
            return False, "Password must be at least 8 characters"
        if not any(c.isupper() for c in password):
            return False, "Password must contain at least one uppercase letter"
        if not any(c.isdigit() for c in password):
            return False, "Password must contain at least one digit"
        if not any(c in "!@#$%^&*()_+-=[]{}|;:,.<>?" for c in password):
            return False, "Password must contain at least one special character"
        return True, "OK"


class JWTManager:
    @staticmethod
    def create_access_token(payload: Dict[str, Any]) -> str:
        data = payload.copy()
        data["exp"] = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
        data["iat"] = datetime.utcnow()
        data["jti"] = str(uuid.uuid4())
        data["type"] = "access"
        return jwt.encode(data, settings.JWT_SECRET_KEY, algorithm=settings.JWT_ALGORITHM)

    @staticmethod
    def create_refresh_token(payload: Dict[str, Any]) -> str:
        data = payload.copy()
        data["exp"] = datetime.utcnow() + timedelta(days=settings.REFRESH_TOKEN_EXPIRE_DAYS)
        data["iat"] = datetime.utcnow()
        data["jti"] = str(uuid.uuid4())
        data["type"] = "refresh"
        return jwt.encode(data, settings.JWT_SECRET_KEY, algorithm=settings.JWT_ALGORITHM)

    @staticmethod
    def decode_token(token: str) -> Dict[str, Any]:
        try:
            payload = jwt.decode(
                token,
                settings.JWT_SECRET_KEY,
                algorithms=[settings.JWT_ALGORITHM],
                options={"verify_exp": True},
            )
            return payload
        except jwt.ExpiredSignatureError:
            raise ValueError("Token has expired")
        except jwt.InvalidTokenError as e:
            raise ValueError(f"Invalid token: {str(e)}")

    @staticmethod
    def hash_refresh_token(token: str) -> str:
        return hashlib.sha256(token.encode()).hexdigest()


class MFAManager:
    @staticmethod
    def generate_secret() -> str:
        return pyotp.random_base32()

    @staticmethod
    def get_totp_uri(secret: str, username: str) -> str:
        totp = pyotp.TOTP(secret)
        return totp.provisioning_uri(
            name=username,
            issuer_name="PharmacyOps"
        )

    @staticmethod
    def verify_totp(secret: str, code: str) -> bool:
        totp = pyotp.TOTP(secret)
        return totp.verify(code, valid_window=1)  # ±30s window


class RateLimiter:
    MAX_LOGIN_ATTEMPTS = 10
    LOCKOUT_DURATION_MINUTES = 15

    @staticmethod
    def should_lock(failed_attempts: int) -> bool:
        return failed_attempts >= RateLimiter.MAX_LOGIN_ATTEMPTS

    @staticmethod
    def get_lockout_until() -> datetime:
        return datetime.utcnow() + timedelta(minutes=RateLimiter.LOCKOUT_DURATION_MINUTES)
