backend/auth_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
python-multipart==0.0.9
redis==5.0.3
httpx==0.27.0
qrcode[pil]==7.4.2
email-validator==2.1.1
